refactor : 공통으로 사용하는 코드 최상위 컴포넌트에 옮기기
