+++ our-app/src/styled/filterTablistStyle.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const TabListCont = styled.section`
+width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 24px 0;
+`
